+++ src/cpu/gdt.c
@@ -6,7 +6,7 @@
 GDT g_gdt[NO_GDT_DESCRIPTORS];
 GDT_PTR g_gdt_ptr;
 
-void gdt_set_entry(int index, uint32 base, uint32 limit, uint8 access, uint8 gran)
+void gdt_set_entry(int index, uint32_t base, uint32_t limit, uint8_t access, uint8_t gran)
 {
     GDT *this = &g_gdt[index];
 
@@ -25,7 +25,7 @@ void gdt_set_entry(int index, uint32 base, uint32 limit, uint8 access, uint8 gra
 void gdt_init()
 {
     g_gdt_ptr.limit = sizeof(g_gdt) - 1;
-    g_gdt_ptr.base_address = (uint32)g_gdt;
+    g_gdt_ptr.base_address = (uintptr_t)g_gdt;
 
     gdt_set_entry(0, 0, 0, 0, 0);
     gdt_set_entry(1, 0, 0xFFFFFFFF, 0x9A, 0xCF);
@@ -33,5 +33,5 @@ void gdt_init()
     gdt_set_entry(3, 0, 0xFFFFFFFF, 0xFA, 0xCF);
     gdt_set_entry(4, 0, 0xFFFFFFFF, 0xF2, 0xCF);
 
-    load_gdt((uint32)&g_gdt_ptr);
+    load_gdt((uint32_t)&g_gdt_ptr);
 }
